requirements-dev.txt
-r requirements.txt

beautifulsoup4==4.9.0
black==22.8.0
bs4==0.0.1
codecov==2.1.12
flake8==4.0.1
ipython==8.5.0
isort==5.10.1
pluggy==1.0.0
pymakefile==0.1.2
pytest-celery==0.0.0
pytest-cov==3.0.0
pytest-django-queries==1.0.0
pytest-django==4.5.2
pytest-forked==1.4.0
pytest-mock==3.8.2
pytest-vcr==1.0.2
pytest-xdist==1.34.0
pytest==7.1.3
python-slugify==1.2.6
